move code to mean linting
