+++ src/vs/workbench/contrib/extensions/common/extensions.ts
@@ -100,8 +100,6 @@ export const AutoUpdateConfigurationKey = 'extensions.autoUpdate';
 export const AutoCheckUpdatesConfigurationKey = 'extensions.autoCheckUpdates';
 export const ShowRecommendationsOnlyOnDemandKey = 'extensions.showRecommendationsOnlyOnDemand';
 export const CloseExtensionDetailsOnViewChangeKey = 'extensions.closeExtensionDetailsOnViewChange';
-// {{SQL CARBON EDIT}}
-export const ExtensionsPolicyKey = 'extensions.extensionsPolicy';
 
 export interface IExtensionsConfiguration {
 	autoUpdate: boolean;
@@ -112,14 +110,6 @@ export interface IExtensionsConfiguration {
 	// {{SQL CARBON EDIT}}
 	extensionsPolicy: string;
 }
-
-// {{SQL CARBON EDIT}}
-export enum ExtensionsPolicy {
-	allowAll = 'allowAll',
-	allowNone = 'allowNone',
-	allowMicrosoft = 'allowMicrosoft'
-}
-// {{SQL CARBON EDIT}} - End
 
 export interface IExtensionContainer {
 	extension: IExtension | null;
